Refactoring Bencode as raw functions instead of method calls to struct.
Lets me more easily repackage this as a library, and since the struct stored no data, it seems unimportant to keep. Since I've also switched to walking an index through the array, instead of reversing the index, popping last entry and passing the array pointer, I could actually use a struct to store the index, removing the need to pass the index every time, I think conceptually it's dumb for the struct to be used like that -> would rather call decode(data: stream) directly.

Would also like to profile walking through an index vs reversing it, popping it, and passing the pointer. I assume Swift optimises the first case and doesn't copy but not entirely sure.